+++ packages/js/src/ui/context/FeedsContext.tsx
@@ -2,40 +2,38 @@ import { Accessor, createContext, ParentComponent, useContext } from 'solid-js';
 import { createStore } from 'solid-js/store';
 import { FetchFeedArgs, Notification } from '../../feeds';
 import { NotificationStatus } from '../../types';
-import { useFeeds } from '../api';
+import { useFeed } from '../api';
 
 type FeedsContextType = {
-  feeds: Accessor<Notification[]>;
+  feed: Accessor<Notification[]>;
   setFeedOptions: (options: FetchFeedArgs) => void;
   feedOptions: FetchFeedArgs;
 };
 
-const FeedsContext = createContext<FeedsContextType | undefined>(undefined);
+const FeedContext = createContext<FeedsContextType | undefined>(undefined);
 
 // TODO: update this after Filter is implemented
-export const FeedsProvider: ParentComponent = (props) => {
+export const FeedProvider: ParentComponent = (props) => {
   const [options, setOptions] = createStore<FetchFeedArgs>({
     status: NotificationStatus.UNREAD,
   });
 
-  const { feeds } = useFeeds({ options });
+  const { feed } = useFeed({ options });
 
   const setFeedOptions = (options: FetchFeedArgs) => {
     setOptions(options);
   };
 
   return (
-    <FeedsContext.Provider value={{ feeds, setFeedOptions, feedOptions: options }}>
-      {props.children}
-    </FeedsContext.Provider>
+    <FeedContext.Provider value={{ feed, setFeedOptions, feedOptions: options }}>{props.children}</FeedContext.Provider>
   );
 };
 
-export const useFeedsContext = () => {
-  const context = useContext(FeedsContext);
+export const useFeedContext = () => {
+  const context = useContext(FeedContext);
 
   if (!context) {
-    throw new Error('useFeedsContext must be used within a FeedsProvider');
+    throw new Error('useFeedContext must be used within a FeedsProvider');
   }
 
   return context;
